requirements.txt
alembic==1.4.2
autopep8==1.5.3
blinker==1.4
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
flake8==3.8.3
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.3
google-auth==1.18.0
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
requests==2.24.0
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.18
toml==0.10.1
urllib3==1.25.9
Werkzeug==1.0.1
zipp==3.1.0
pymysql==0.9.3
